[Fix] : 스킬 밸런스 조정(하향) 및 틱데미지 버그 수정(크리티컬 계산시 누락됨), 플레이어 스탯 로직 수정

+++ Assets/02.Scripts/02.Data/00.Loader/StatUpgradeDB.cs
@@ -24,7 +24,7 @@ public class StatUpgradeDB
     /// <summary>
     /// MaxHealthBaseStat
     /// </summary>
-    public int MaxHealthBaseStat;
+    public float MaxHealthBaseStat;
 
     /// <summary>
     /// AtkBaseCost
@@ -39,7 +39,7 @@ public class StatUpgradeDB
     /// <summary>
     /// AtkBaseStat
     /// </summary>
-    public int AtkBaseStat;
+    public float AtkBaseStat;
 
     /// <summary>
     /// DefBaseCost
@@ -54,7 +54,7 @@ public class StatUpgradeDB
     /// <summary>
     /// DefBaseStat
     /// </summary>
-    public int DefBaseStat;
+    public float DefBaseStat;
 
     /// <summary>
     /// ReduceDamageBaseCost
@@ -69,7 +69,7 @@ public class StatUpgradeDB
     /// <summary>
     /// ReduceDamageBaseStat
     /// </summary>
-    public int ReduceDamageBaseStat;
+    public float ReduceDamageBaseStat;
 
     /// <summary>
     /// CriticalRateBaseCost
@@ -84,7 +84,7 @@ public class StatUpgradeDB
     /// <summary>
     /// CriticalRateBaseStat
     /// </summary>
-    public int CriticalRateBaseStat;
+    public float CriticalRateBaseStat;
 
     /// <summary>
     /// CriticalDamageBaseCost
@@ -99,7 +99,7 @@ public class StatUpgradeDB
     /// <summary>
     /// CriticalDamageBaseStat
     /// </summary>
-    public int CriticalDamageBaseStat;
+    public float CriticalDamageBaseStat;
 
 }
 public class StatUpgradeDBLoader
